itest: rename to integration-test everywhere
Rename the itest/ directory to integration-test/ for clarity. Update
all references: Go module (pyroscope-dotnet-integration-test), package
name (integrationtest), Dockerfiles (integration-test.Dockerfile,
integration-test-with-otel.Dockerfile), CI workflow paths and job
names, and Docker image tag suffixes.
