+++ integration-test-with-otel.Dockerfile
